Introduce msp scope
This eventually and this codifies internal workflows now instead of
requiring changes later.

This creates a new type used for scopes that allows the ID to be `msp`
and associated changes to tables, but also adds restrictions ensuring
that we can only ever have one `msp` scope and that we are not allowed
to delete it. Tests have been added or updated.