Add grant scope ID to roles (#165)
Add grant scope ID to roles.

This allows specifying a scope for the grants that is a child of the
current scope.

Checking is implemented as a database trigger/procedure with an
appropriate not-null constraint. We set it to the role's scope ID if not
provided or empty. Unlike the role's scope_id this can change later so
long as it satisfies the checks.