Inclusion of the error description
Is your feature request related to a problem? Please describe.
During a failure in creating the target, the exception caused did not come with a description capable of bringing a quick understanding of the problem, just the stacktrace.

Describe the solution you'd like
Use the existing module to include messages.