Add a callback mechanism
This mechanism intercepts any request with the path ending in
`:callback`.

Currently this is limited to GET (but can support others), but in this
case will parse the form, and put the resulting variables in a
predictable format that is then sent along as JSON to grpc-gateway for
normal handling.