Add a Docker entrypoint script and modify Dockerfiles to mimic Vault, to enable mlock.
* Add entrypoint script with logic borrowed from Vault's.  This script adds
IPC_LOCK capability at runtime and then executes Boundary under its own user.

* Modify Dockerfiles to use the entrypoint script.

* Remove USER directives from the Boundary Dockerfiles.  This is required to
allow the entrypoint to run setcap.

Because Boundary itself is no longer the entrypoint, the behavior of the
container with respect to passing or overriding arguments is different now.
The Vault script's logic around handling this was not copied entirely -- only
to the extent required to allow the `server` subcommand to run properly.