Add a test for existing GrantsForUser functionality.
The test has randomization built in to allocate users across groups and
both users and groups across roles. It tracks which users and groups
have been assigned to which roles, then ensures that GrantsForUser
returns the correct set of roles for a given user.

This required modifying the CTE in GrantsForUser to return role_id as
well.

This doesn't test managed groups yet, that will come in a follow-on PR.