Remove mock worker testing
This removes the mock worker "unit" testing (for now).

We may re-add this in the future - but for now we've decided that the
tests are not 100% necessary and that the E2E tests are good enough.

Some time will be taken to re-factor and see how we can formalize some
of the things done in the test to help with readability, re-usability,
etc.