test(db): Thin line for new sql testing suite
This sets up the configuration for running a new suite of tests that are
written in sql, to test the sql logic implemented in plpgsql functions
and triggers. It is primarily designed to provides tests for the data
warehouse functionality that is implemented purely in sql.

The infrasturture to run these tests leverages docker. It creates one
docker container using the `library/postgres` image and provisions the
database with the boundary sql migrations and test helper functions.
A separate docker container is run that has the pgtap extension and
pg_prove cli to execute the tests.

As noted, pgtap is used for the tests. This postgresql extension
provides a number of helpful assertion functions to make writting sql
tests easier. It also, along with the pg_prove cli, provides readable
test output.

This commit also provides an initial test for the wh_user_dimension to
establish an inital pattern for these kind of tests.

Lastly this configures circleci to run this test suite as a separate job
in parallel with the other tests.

See:
    https://pgtap.org/