Add address field to API host message
This adds a top-level, read-only "address" field to the host API
message. This field will be used to populate the host address for
plugin-based hosts when their respective host sets are queried.
Attributes will be populated with plugin-specific data that does not
include address data.

Note that this does not affect how static addresses are created or
updated - the API still expects the address in attributes. This keeps
the CRUD workflow consistent with the new plugin-based host catalog and
set types, which take their configuration data in attributes as well.

A small CLI update has been added to output the address at the top-level
as well.